+++ Assets/Systems/Attack System/Scripts/SpawnCapsule.cs
@@ -15,7 +15,7 @@ public class SpawnCapsule : MonoBehaviour
         capsuleLocations = new GameObject[GameObject.Find("AttackAreas").transform.childCount];
     }
 
-    public void Activate(Transform spawnPosition, int location, int offsetx, int offsety, int offsetz, float capsuleRotation)
+    public void Activate(Transform spawnPosition, int location, float offsetx, float offsety, float offsetz, float capsuleRotation)
     {
         capsuleLocations[location] = Instantiate(playerCapsule, new Vector3(spawnPosition.transform.position.x + offsetx, spawnPosition.transform.position.y + offsety, spawnPosition.transform.position.z + offsetz), Quaternion.identity, spawnPosition);
         capsuleLocations[location].transform.localEulerAngles = new Vector3(0, capsuleRotation, 0);
